Move stdlib/surface to stdlib/area/surface
